+++ src/types/utils.ts
@@ -46,3 +46,9 @@ export type GetOddOrEvenSegments<
 	: T extends `${string}/${infer R extends string}`
 	? GetOddOrEvenSegments<R, true, ACC>
 	: ACC
+
+export type DeepPartial<T> = {
+	[K in keyof T]?: T[K] extends Record<string, unknown>
+		? DeepPartial<T[K]>
+		: T[K]
+}